Added SchoolMember class and implemented inheritance for Instructor/Student classes.
